KW-CAI-HS-T8-main/requirements.txt
streamlit
torch
Pillow
textblob
transformers
google-generativeai